+++ src/api/projects/create.rs
@@ -541,7 +541,7 @@ pub struct CreateProject<'a> {
     operations_access_level: Option<FeatureAccessLevel>,
     /// Set the access level for requirements features.
     #[builder(default)]
-    requirements_access_level: Option<FeatureAccessLevelPublic>,
+    requirements_access_level: Option<FeatureAccessLevel>,
     /// Set the access level for analytics features.
     #[builder(default)]
     analytics_access_level: Option<FeatureAccessLevel>,
@@ -1492,14 +1492,14 @@ mod tests {
             .method(Method::POST)
             .endpoint("projects")
             .content_type("application/x-www-form-urlencoded")
-            .body_str(concat!("name=name", "&requirements_access_level=public"))
+            .body_str(concat!("name=name", "&requirements_access_level=enabled"))
             .build()
             .unwrap();
         let client = SingleTestClient::new_raw(endpoint, "");
 
         let endpoint = CreateProject::builder()
             .name("name")
-            .requirements_access_level(FeatureAccessLevelPublic::Public)
+            .requirements_access_level(FeatureAccessLevel::Enabled)
             .build()
             .unwrap();
         api::ignore(endpoint).query(&client).unwrap();